feat(plugins): add plugins support to cortex devices
Added a new 'Plugins' tab in the Cortex Devices interface. This includes changes to the HTML and TypeScript files for Cortex Devices, as well as creation of new JSON data files for Plini and Gojira plugins. Also added a new Plugins Table view with corresponding HTML and TypeScript files. Updated the AllTable class to filter by deviceType instead of type.